pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "parc-ferme"
dynamic = ["version"]
description = "Automated PR code reviews using Claude AI"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Passakorn"},
]
readme = "README.md"
keywords = ["code-review", "github", "claude", "cli", "pull-request"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/passakorn/parc-ferme"
Repository = "https://github.com/passakorn/parc-ferme"
Issues = "https://github.com/passakorn/parc-ferme/issues"

[project.scripts]
parc-ferme = "parc_ferme.cli:main"
parc = "parc_ferme.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.dynamic]
version = {attr = "parc_ferme.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
